tools/green_days/requirements.txt
pyyaml==6.0
pytz==2021.3
